Upgrade mathf lib , update float from float32=>float64 to match gopher's habits

+++ cmd/codegen/generate/common/funcs.go
@@ -429,7 +429,7 @@ func GetManagers(ast clang.CHeaderFileAST) []string {
 	sort.Strings(managers)
 	cppType2Go = map[string]string{
 		"GdInt":    "int64",
-		"GdFloat":  "float32",
+		"GdFloat":  "float64",
 		"GdObj":    "Object",
 		"GdVec2":   "Vec2",
 		"GdVec3":   "Vec3",
